requirements.txt
asgiref==3.5.2
<<<<<<< HEAD
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
distlib==0.3.6
Django==4.1.2
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.8.0
Flask==1.1.2
funcparserlib==0.3.6
huggingface-hub==0.10.1
hypothesis==6.0.0
idna==2.10
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mido==1.2.9
nltk==3.7
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
Pillow==8.1.0
pipenv==2022.10.12
platformdirs==2.5.2
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pygame==2.0.1
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-ta==1.6.3
pytz==2021.1
PyYAML==6.0
regex==2022.9.13
requests==2.25.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tokenizers==0.13.1
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.24.0
trollius==2.1.post2
twilio==6.58.0
typeguard==2.10.0
typing_extensions==4.3.0
urllib3==1.26.2
virtualenv==20.16.5
virtualenv-clone==0.5.7
Werkzeug==1.0.1
wrapt==1.12.1
=======
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.1
django-recurrence==1.11.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
>>>>>>> 2a2cd131ab18be77d78d7d87195a421d721850a4
